+++ src/main/kotlin/frontend/helpers/Extensions.kt
@@ -15,4 +15,4 @@ class Extensions {
     }
 }
 
-fun String.toPrice() = filter { it.isDigit() }.toFloat() / 100f
+fun String.toPrice() = filter { it.isDigit() }.toDouble() / 100
